requirements.txt
# Multi-Agent System Requirements

# Core Framework
fastapi==0.122.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20

# Google Gemini AI
google-generativeai==0.8.5

# Data Processing
pandas==2.3.3
numpy==2.2.6

# Utilities
pydantic==2.12.5
python-dotenv==1.2.1

